% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alignment.R
\name{Alignment}
\alias{Alignment}
\alias{countCharFreq}
\title{Alignment analyser}
\usage{
countCharFreq(
  seqs.df,
  attr.taxon = "taxon",
  attr.seq = "value",
  CASE.FUN = toupper
)
}
\arguments{
\item{seqs.df}{the \code{\link{tibble}} returned by \code{\link{getSeqsDF}}.}

\item{attr.taxon, attr.seq}{attribute names required in BEAST 2 XML.}

\item{CASE.FUN}{the function (to use upper case as default).}
}
\description{
Analyse the alignment extracted from BEAST 2 XML.
The input is \code{\link{tibble}}.
}
\details{
\code{countCharFreq} splits sequences and counts alphabet frequency.
}
\examples{
seqs.df <- getSeqsDF(xml)
freq <- countCharFreq(seqs.df, CASE.FUN=tolower)
print(freq, n=Inf)

}
\keyword{alignment}
