% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{Metadata}
\alias{Metadata}
\alias{extractMeta}
\title{Metadata}
\usage{
extractMeta(
  seqs.df,
  attr.taxon = "taxon",
  sep = "\\\\|",
  cols = c("id", "date", "country", "city"),
  ...
)
}
\arguments{
\item{seqs.df}{The \code{\link{tibble}} returned by \code{\link{getSeqsDF}}.}

\item{attr.taxon}{Attribute for a taxon name in BEAST 2 XML.}

\item{sep}{Separator between columns, see \code{\link{separate}}.}

\item{cols}{Column names after splitting \code{taxon} into multiple columns.}

\item{...}{Other arguments passed to \code{\link{separate}}.}
}
\description{
Analyse the metadata extracted from taxa names in the BEAST 2 XML.
The input is \code{\link{tibble}}.
}
\details{
\code{extractMeta} extracts meta data from taxa names into a \code{\link{tibble}}.
For example,
\code{<sequence id="seq1" spec="Sequence" taxon="seq1|2020-04-25|New_Zealand|Auckland"
totalcount="4" value="acgt"/>}.
}
\examples{
seqs.df <- getSeqsDF(xml)
# seq1|2020-04-25|New_Zealand|Auckland
meta <- extractMeta(seqs.df, sep="\\\\|", cols = c("id","date","country","city"))
meta <- meta \%>\% mutate(month=gsub("(\\\\d+)-(\\\\d+)-.*","\\\\1-\\\\2",date))
print(meta, n=Inf)
write_delim(meta, "meta.tsv", delim = "\t")

}
\keyword{metadata}
