% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataPlot.R, R/PlotDate.R
\name{Plot}
\alias{Plot}
\alias{plotLocDate}
\alias{barLocDate}
\title{Visualisations}
\usage{
plotLocDate(
  meta.df,
  date.col = "date",
  trait.col = "trait",
  x.lab = "Date",
  y.lab = "Number of samples",
  legend = "locations"
)

barLocDate(
  meta,
  x = "date",
  fill = "region",
  x.lab.format = date_format("\%m-\%Y"),
  title = "",
  xlab = "Date",
  ylab = "Number of samples",
  ...
)
}
\arguments{
\item{meta.df}{data frame to contain samples, dates and locations.}

\item{meta}{The \code{\link{tibble}} returned by \code{\link{extractMeta}}.}

\item{x, fill}{Data column for x axis and colour.}

\item{x.lab.format}{x labels, default to month.}

\item{title, xlab, ylab}{title, x and y axis labels.}

\item{...}{Other arguments passed to \code{\link{geom_bar}}.}
}
\description{
Visualisations using \code{\link{ggplot2}}.

Visulizations using \code{\link{ggplot2}}. The input is \code{\link{tibble}}.
}
\details{
\code{plotLocDate} plots samples through time coloured by traits (locations).

\code{barLocDate} plots a bar chart to count the number of samples
through the sampling date, and coloured by locations.
}
\examples{
dates.df <- getDatesDF(xml, xpath="//trait")
loc.df <- getLocDF(xml, xpath="//traitSet")
meta.df <- inner_join(dates.df, loc.df, by = "taxon")
stopifnot(nrow(meta.df)==nrow(dates.df) && nrow(meta.df)==nrow(loc.df))
plotLocDate(meta.df)

meta <- extractMeta(seqs.df, sep="\\\\|", cols = c("id","date","country","city"))
# need both sampling date and location
p <- barLocDate(meta, fill = "country")
title=paste(nrow(meta),"samples from", length(unique(meta[["region"]])),
      "regions from between", date_format("\%d-\%m-\%Y")(min(meta[["date"]])),
      "and", date_format("\%d-\%m-\%Y")(max(meta[["date"]])))

}
\keyword{plot}
