% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLEditor.R
\name{XMLEditor}
\alias{XMLEditor}
\alias{replaceSeqs}
\alias{replaceChainLength}
\alias{replaceStoreEvery}
\alias{replaceLogEvery}
\title{edit BEAST 2 XML}
\usage{
replaceSeqs(xml, node.seqs, xpath = "//data")

replaceChainLength(xml, value, attr = "chainLength", xpath = "//run")

replaceStoreEvery(xml, value, attr = "storeEvery", xpath = "//state")

replaceLogEvery(
  xml,
  ids = c("tracelog", "treelog", "screenlog"),
  values = c("10000", "10000", "1000000"),
  attr = "logEvery",
  xpath = "//logger"
)
}
\arguments{
\item{ids}{a vetor of id for logging.}

\item{values}{a vetor of values for these ids. which must have the same length.}
}
\description{
The functions to edit BEAST 2 XML using \code{\link{xml2}}.
}
\details{
\code{replaceSeqs} replaces the old sequences in XML <data> as the given new sequences.

\code{replaceChainLength} replaces the chain length in MCMC.

\code{replaceStoreEvery} replaces the store frequency in MCMC.

\code{replaceLogEvery} replaces the logging frequency in MCMC.
}
\examples{
xml <- read_xml(file.path(DATA.XML))
SEQS <- xml_find_all(xml, "//sequence")
template <- read_xml(file.path(TMPL.XML))
# replace data in TMPL.XML by data in DATA.XML
replaceSeqs(template, SEQS)

}
\keyword{xml}
